requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
CacheControl==0.14.2
cachetools==5.5.2
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
factory_boy==3.3.3
Faker==37.1.0
firebase-admin==6.7.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
kombu==5.5.3
mpmath==1.3.0
msgpack==1.1.0
mysqlclient==2.2.7
numpy==2.2.4
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pluggy==1.5.0
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
threadpoolctl==3.6.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
